Added the following:
1. Counterbalance the response keys for A/B & A/-A (Summerfield says
within and between conditions)
3. Make the line for A/-A to be green.
4. Put a circle filled with grey colour, on which the orientation lines
will be placed in instructions, as in Summerfield.
5. Make sure that As and Bs (as well as A and -As) are 50-50 on average.

Finally, a couple of things I don't understand:
a. In the instructions' window in Psychopy you have chosen for fRed to
be ''constant'', while for jBlue to be ''set every repeat'' (it's
actually a variable). Also, there is a difference as to opacity between
the lineRed and lineBlue properties.
b. I don't know whether the signs f=red and j=blue should in some way
be attached to the correct colour because the following example I met
may be quite misleading for participants.